@@ -1,10 +1,12 @@
+'use client'
+
 import { GridPattern } from '@/components/GridPattern'
 
 export function HeroPattern() {
   return (
     <div className="absolute inset-0 -z-10 mx-0 max-w-none overflow-hidden">
       <div className="absolute left-1/2 top-0 ml-[-38rem] h-[25rem] w-[81.25rem] dark:[mask-image:linear-gradient(white,transparent)]">
-        <div className="absolute inset-0 bg-gradient-to-r from-indigo-200 to-indigo-800 opacity-40 [mask-image:radial-gradient(farthest-side_at_top,white,transparent)] dark:from-[#36b49f]/30 dark:to-[#DBFF75]/30 dark:opacity-100">
+        <div className="absolute inset-0 bg-gradient-to-r from-purple-200 to-blue-600 opacity-40 [mask-image:radial-gradient(farthest-side_at_top,white,transparent)] dark:from-[#4A28FF]/30 dark:to-[#9ED6FF]/30 dark:opacity-100">
           <GridPattern
             width={72}
             height={56}
